requirements.txt
opencv-python
pyzbar
numpy
qrcode[pil]
flask
pillow